Move accounting tests out of runtests.c
